add permute groovy example, add permute docs
