+++ webtau-core-groovy/src/test/groovy/com/twosigma/webtau/data/table/TableDataExtensionTest.groovy
@@ -18,6 +18,8 @@ package com.twosigma.webtau.data.table
 
 import org.junit.Test
 
+import static com.twosigma.webtau.Ddjt.*
+
 class TableDataExtensionTest {
     @Test
     void "should register header and values using pipes"() {
@@ -44,6 +46,19 @@ class TableDataExtensionTest {
     }
 
     @Test
+    void "should generate multiple rows from multi-values"() {
+        def tableData = createTableDataWithPermute()
+
+        assert tableData.numberOfRows() == 6
+        assert tableData.row(0).toMap() == ['Col A': true, 'Col B': 'v1b', 'Col C': 'a']
+        assert tableData.row(1).toMap() == ['Col A': false, 'Col B': 'v1b', 'Col C': 'a']
+        assert tableData.row(2).toMap() == ['Col A': true, 'Col B': 'v1b', 'Col C': 'b']
+        assert tableData.row(3).toMap() == ['Col A': false, 'Col B': 'v1b', 'Col C': 'b']
+        assert tableData.row(4).toMap() == ['Col A': 'v2a', 'Col B': 10, 'Col C': 'v2c']
+        assert tableData.row(5).toMap() == ['Col A': 'v2a', 'Col B': 20, 'Col C': 'v2c']
+    }
+
+    @Test
     void "should ignore underscore under header"() {
         def table = ["hello" | "world"] {
                     ___________________
@@ -63,6 +78,13 @@ class TableDataExtensionTest {
         ["Col A" | "Col B" | "Col C"] {
            "v1a" |   "v1b" | "v1c"
            "v2a" |   "v2b" | "v2c" }
+    }
+
+    static TableData createTableDataWithPermute() {
+        ["Col A"              | "Col B"         | "Col C"] {
+        ___________________________________________________________
+         permute(true, false) | "v1b"           | permute('a', 'b')
+         "v2a"                | permute(10, 20) | "v2c" }
     }
 
     private static void validateTableData(TableData tableData) {
