+++ webtau-core/src/test/groovy/com/twosigma/webtau/data/table/TableDataTest.groovy
@@ -16,6 +16,7 @@
 
 package com.twosigma.webtau.data.table
 
+import com.twosigma.webtau.documentation.DocumentationArtifacts
 import org.junit.Test
 
 import static com.twosigma.webtau.Ddjt.*
@@ -41,10 +42,8 @@ class TableDataTest {
     }
 
     @Test
-    void "should generate multiple raws from multi-values"() {
-        def tableData = table("Col A", "Col B", "Col C").values(
-                 permute(true, false), "v1b", permute('a', 'b'),
-                                "v2a", permute(10, 20), "v2c")
+    void "should generate multiple rows from multi-values"() {
+        def tableData = createTableDataWithPermute()
 
         assert tableData.numberOfRows() == 6
         assert tableData.row(0).toMap() == ['Col A': true, 'Col B': 'v1b', 'Col C': 'a']
@@ -53,6 +52,8 @@ class TableDataTest {
         assert tableData.row(3).toMap() == ['Col A': false, 'Col B': 'v1b', 'Col C': 'b']
         assert tableData.row(4).toMap() == ['Col A': 'v2a', 'Col B': 10, 'Col C': 'v2c']
         assert tableData.row(5).toMap() == ['Col A': 'v2a', 'Col B': 20, 'Col C': 'v2c']
+
+        DocumentationArtifacts.create(TableDataTest, 'table-with-permute.json', tableData.toJson())
     }
 
     static TableData createTableDataSeparateValues() {
@@ -63,9 +64,16 @@ class TableDataTest {
 
     static TableData createTableDataInOneGo() {
         table("Col A", "Col B", "Col C",
-                ________________________________,
+              ________________________________,
                 "v1a",   "v1b", "v1c",
                 "v2a",   "v2b", "v2c")
+    }
+
+    static TableData createTableDataWithPermute() {
+        table("Col A"              , "Col B"         , "Col C",
+              ________________________________________________________________,
+               permute(true, false), "v1b"           , permute('a', 'b'),
+               "v2a"               , permute(10, 20) , "v2c")
     }
 
     private static void validateTableData(TableData tableData) {
