+++ webtau-http/src/main/java/com/twosigma/webtau/http/HttpDocumentation.java
@@ -16,7 +16,7 @@
 
 package com.twosigma.webtau.http;
 
-import com.twosigma.documentation.DocumentationArtifactsLocation;
+import com.twosigma.webtau.documentation.DocumentationArtifactsLocation;
 import com.twosigma.webtau.http.validation.HttpValidationResult;
 import com.twosigma.webtau.utils.FileUtils;
 import com.twosigma.webtau.utils.JsonUtils;
